+++ Source/EventFlow.Tests/UnitTests/Extensions/EventUpgraderExtensionsTests.cs
@@ -20,21 +20,43 @@
 // COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// 
+//
 
-using System.ComponentModel;
-using EventFlow.Configuration;
-using EventFlow.TestHelpers;
-using EventFlow.TestHelpers.Suites;
+using System;
+using System.Collections.Generic;
+using EventFlow.Aggregates;
+using EventFlow.Configuration.Registrations;
+using EventFlow.EventStores;
+using EventFlow.Extensions;
+using EventFlow.TestHelpers.Aggregates;
+using FluentAssertions;
+using NUnit.Framework;
 
-namespace EventFlow.Tests.IntegrationTests.EventStores
+namespace EventFlow.Tests.UnitTests.Extensions
 {
-    [Category(Categories.Integration)]
-    public class InMemoryEventStoreTests : TestSuiteForEventStore
+    public class EventUpgraderExtensionsTests
     {
-        protected override IRootResolver CreateRootResolver(IEventFlowOptions eventFlowOptions)
+        [Test]
+        public void AbstractEventUpgraderIsNotRegistered()
         {
-            return eventFlowOptions.CreateResolver();
+            // Arrange
+            var registry = new AutofacServiceRegistration();
+            var sut = EventFlowOptions.New.UseServiceRegistration(registry);
+
+            // Act
+            Action act = () => sut.AddEventUpgraders(new List<Type>
+            {
+                typeof (AbstractTestEventUpgrader)
+            });
+
+            // Assert
+            act.ShouldNotThrow<ArgumentException>();
         }
+    }
+
+    public abstract class AbstractTestEventUpgrader : IEventUpgrader<ThingyAggregate, ThingyId>
+    {
+        public abstract IEnumerable<IDomainEvent<ThingyAggregate, ThingyId>> Upgrade(
+            IDomainEvent<ThingyAggregate, ThingyId> domainEvent);
     }
 }